hand_controller: inline bin/ helpers into hc2keelson.py
The all-in-one Dockerfile copies every connector's bin/*.py flat into
/usr/local/bin/ -- helper modules with generic names (terminal_inputs.py)
would silently overwrite each other across connectors with no build-time
warning. Until a repo-wide isolation scheme lands, hand_controller keeps
nothing but a single self-contained entry point in bin/.

- Inlined joystick_proto.py constants + read_event + normalize_axis.
- Inlined the terminal_inputs() argparse function.
- Deleted bin/joystick_proto.py and bin/terminal_inputs.py.
- examples/joystick_reader.py now ships its own copy of the 3 constants
  and 2 functions (matches its docstring promise of no external libs).
- Dropped sys.path.insert in tests/conftest.py -- not needed any more.
- CLAUDE.md documents why no sibling modules in bin/.
