+++ connectors/hand_controller/bin/hc2keelson.py
@@ -14,6 +14,7 @@ Publishes axes and buttons as TimestampedInt messages. Unmapped buttons
 publish to button_pressed/button_released with the button number as value.
 """
 
+import argparse
 import sys
 import os
 import json
@@ -34,14 +35,45 @@ import keelson
 from keelson.payloads.Primitives_pb2 import TimestampedInt, TimestampedFloat
 from keelson.scaffolding.liveliness import declare_liveliness_token
 
-from terminal_inputs import terminal_inputs
-from joystick_proto import (
-    JS_EVENT_BUTTON,
-    JS_EVENT_AXIS,
-    JS_EVENT_INIT,
-    read_event,
-    normalize_axis,
-)
+
+# --- Linux joystick HID event protocol -------------------------------------
+# Inlined here (rather than in a sibling module) so this entry-point script
+# is the only file that lands in /usr/local/bin/ in the container image —
+# avoids the silent-collision risk of flat helper modules with generic names
+# (see connectors/hand_controller/CLAUDE.md).
+
+# Linux joystick event types (from linux/joystick.h)
+JS_EVENT_BUTTON = 0x01  # Button pressed/released
+JS_EVENT_AXIS = 0x02  # Joystick moved
+JS_EVENT_INIT = 0x80  # Initial state flag (OR'd with the real type)
+
+# 8-byte event format: timestamp (uint32), value (int16), type (uint8), number (uint8)
+_EVENT_FORMAT = "IhBB"
+_EVENT_SIZE = 8
+
+
+def read_event(device_file):
+    """Read a single 8-byte joystick event from device_file.
+
+    Returns (timestamp, value, event_type, number), or None on a short read
+    or read error. Callers can treat None uniformly as "no event this tick".
+    """
+    try:
+        data = device_file.read(_EVENT_SIZE)
+    except OSError:
+        return None
+    if len(data) < _EVENT_SIZE:
+        return None
+    return struct.unpack(_EVENT_FORMAT, data)
+
+
+def normalize_axis(value):
+    """Normalize int16 axis value to percent (-100.0..100.0).
+
+    Uses a 32768 divisor for symmetric mapping; clamps as defense in depth.
+    """
+    return max(-100.0, min(100.0, value * 100.0 / 32768.0))
+
 
 # Global state
 PUBLISHERS: Dict[tuple, Any] = {}  # Cache for lazy publisher creation
@@ -417,6 +449,122 @@ def event_source_tcp(host: str, port: int, max_retries: int = 0):
         if not shutdown_requested:
             time.sleep(backoff_seconds)
             backoff_seconds = min(backoff_seconds * 2, max_backoff)
+
+
+def terminal_inputs():
+    """Parse the terminal inputs and return the arguments."""
+    parser = argparse.ArgumentParser(
+        prog="keelson_connector_hc",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+        description="Read joystick/gamepad controller and publish to Keelson/Zenoh",
+    )
+
+    parser.add_argument(
+        "-l",
+        "--log-level",
+        type=int,
+        default=20,
+        help="Log level 10=DEBUG, 20=INFO, 30=WARN, 40=ERROR, 50=CRITICAL 0=NOTSET",
+    )
+    parser.add_argument(
+        "--mode",
+        "-m",
+        dest="mode",
+        choices=["peer", "client"],
+        type=str,
+        help="The zenoh session mode.",
+    )
+    parser.add_argument(
+        "--connect",
+        action="append",
+        type=str,
+        help="Endpoints to connect to, in case multicast is not working. ex. tcp/localhost:7447",
+    )
+    parser.add_argument(
+        "-r",
+        "--realm",
+        default="rise",
+        type=str,
+        help="Unique id for a domain/realm to connect ex. rise",
+    )
+    parser.add_argument(
+        "-e",
+        "--entity-id",
+        default="rov",
+        type=str,
+        help="Entity being a unique id representing an entity within the realm",
+    )
+    parser.add_argument(
+        "--device",
+        "-d",
+        type=str,
+        default="/dev/input/js0",
+        help="Joystick device path (default: /dev/input/js0)",
+    )
+    parser.add_argument(
+        "--relay",
+        type=str,
+        default=None,
+        help="TCP relay address (host:port) for cross-platform mode. "
+        "Reads joystick events from a TCP relay instead of a device file. "
+        "Example: --relay host.docker.internal:9090",
+    )
+    parser.add_argument(
+        "-c",
+        "--controller",
+        type=str,
+        default="ssrov",
+        help="Controller profile name (resolves to profiles/<name>.yaml).",
+    )
+    parser.add_argument(
+        "--controller-config",
+        type=str,
+        default=None,
+        help="Path to a custom controller-profile YAML file. Overrides --controller.",
+    )
+    parser.add_argument(
+        "--relay-max-retries",
+        type=int,
+        default=0,
+        help="Max relay connection attempts before exit (0 = unlimited).",
+    )
+    parser.add_argument(
+        "--axis-min-interval-ms",
+        type=int,
+        default=30,
+        help="Per-axis rate limit: minimum ms between publishes when value barely changed.",
+    )
+    parser.add_argument(
+        "--axis-min-change",
+        type=float,
+        default=1.0,
+        help="Per-axis rate limit: percentage-point change that forces immediate publish.",
+    )
+    parser.add_argument(
+        "--axis-center-snap-pct",
+        type=float,
+        default=0.0,
+        help=(
+            "Snap |value| < this to 0.0 before the rate-limit check. "
+            "Cleans up joystick ADC rest offset so a released stick publishes exactly 0.0. "
+            "0 disables; recommended 2.0. Loses sub-snap precision near rest."
+        ),
+    )
+    parser.add_argument(
+        "--log-json",
+        action="store_true",
+        help="Emit logs as one JSON object per line (for container log pipelines).",
+    )
+    parser.add_argument(
+        "--source-id",
+        type=str,
+        default=None,
+        help="Override the source-id base. Defaults to the --controller value. "
+        "Use this to run two of the same controller side-by-side with distinct "
+        "source-id prefixes (e.g. --controller ssrov --source-id ssrov-port).",
+    )
+
+    return parser.parse_args()
 
 
 def main():
